add InfoButton and ShareButtons components
